backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Chris Spann <cspanndev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
fastapi = "^0.97.0"
fastapi-sqlalchemy = "^0.2.1"
uvicorn = "^0.22.0"
pydantic = "^1.10.9"
sqlalchemy = "^2.0.16"
psycopg2 = "^2.9.6"
alembic = "^1.11.1"
httpx = "^0.24.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
black = "^23.3.0"
ruff = "^0.0.272"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "-ra -q -s -v --cov=. --cov-report term-missing"

[tool.coverage.run]
omit = ["/tests/*"]

[tool.pyright]
reportGeneralTypeIssues=false

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "PL", "SIM", "C", "RET", "N", "UP", "PT"]
ignore = ["PLR2004"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

line-length=100
